Refactor test helper to utils package
Refs #11
